requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gitpython>=3.1.40
requests>=2.31.0
